% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_scenarios.R
\docType{data}
\name{ag_scenarios}
\alias{ag_scenarios}
\title{ag scenarios}
\format{
A vector with 13 elements:
\describe{
  \item{1}{a}
  \item{2}{b}
  \item{3}{d}
  \item{4}{e}
  \item{5}{S1}
  \item{6}{S1_P1}
  \item{7}{S1_P2}
  \item{8}{S2}
  \item{9}{S2_P1}
  \item{10}{S2_P2}
  \item{11}{S3}
  \item{12}{S3_P1}
  \item{13}{S3_P2}
}
}
\usage{
ag_scenarios
}
\description{
agriculture scenarios considered by cbwg
}
\examples{
ag_scenarios

}
\keyword{datasets}
