% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{cbwg_scen_to_fair}
\alias{cbwg_scen_to_fair}
\title{cbwg_scen_to_fair}
\usage{
cbwg_scen_to_fair(
  cbwg_emissions,
  tim_scen = "400mt",
  lulucf_scen = "L4",
  ag_scen = "e"
)
}
\arguments{
\item{cbwg_emissions}{input emissions in kt gas}

\item{tim_scen}{ffi co2 scenario}

\item{lulucf_scen}{lulucf scenario}

\item{ag_scen}{ag scenario}
}
\description{
re-formats CBWG emissions input file scenario to FaIR format and units
}
