% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{cbwg_to_loo_global_emissions}
\alias{cbwg_to_loo_global_emissions}
\title{cbwg_to_loo_global_emissions}
\usage{
cbwg_to_loo_global_emissions(global_scenarios, tim_scenarios, goblin_scenarios)
}
\arguments{
\item{global_scenarios}{a vector SSP scenarios}

\item{tim_scenarios}{vector of Energy system scenarios}

\item{goblin_scenarios}{vector AFOLU scenarios}
}
\value{
emissions
}
\description{
Creates a global emissions dataset for specified global and Irish emissions scenarios in FaIR format. Contains 21 climate forcers.
}
