% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_cbwg_2.R
\docType{data}
\name{emissions_cbwg_2}
\alias{emissions_cbwg_2}
\title{Irish climate forcing emissions 1745-2100}
\format{
A data frame with 12,574 rows and 5 columns:
\describe{
  \item{scenario}{data source prior to 2022, cbwg scenario or other data source after 2022}
  \item{year}{year}
  \item{value}{annual emissions amount}
  \item{gas}{forcing agent}
  \item{units}{units}
}
}
\source{
<https://zenodo.org/records/10904361, CBWG, EPA,.....>
}
\usage{
emissions_cbwg_2
}
\description{
Historical estimates and future scenario data for 23 climate forcing gases relevant to Ireland. Historical data for 1990-2022 are form the National Inventory Report.
Data prior to 1990 are etsimates from a range of sources. Future scenario data for CO2, N2O and CH4 are from second iteration CBWG (105 scenarions)
Scenario labels prior to 2022 indicate the source of the data.
}
\keyword{datasets}
