% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapri_scenarios.R
\docType{data}
\name{fapri_scenarios}
\alias{fapri_scenarios}
\title{fapri scenarios}
\format{
A vector with 9 elements:
\describe{
  \item{1}{S1}
  \item{2}{S1_P1}
  \item{3}{S1_P2}
  \item{4}{S2}
  \item{5}{S2_P1}
  \item{6}{S2_P2}
  \item{7}{S3}
  \item{8}{S3_P1}
  \item{9}{S3_P2}
}
}
\usage{
fapri_scenarios
}
\description{
fapri agriculture scenarios considered by cbwg
}
\examples{
fapri_scenarios

}
\keyword{datasets}
