% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{gen_fair}
\alias{gen_fair}
\title{gen_fair}
\usage{
gen_fair(
  scenarios = global_scenarios,
  endyear = 2100L,
  ch4_meth = "leach2021",
  ghg_meth = "meinshausen2020"
)
}
\arguments{
\item{scenarios}{the global scenarios to be used}

\item{endyear}{final year of calculation}

\item{ch4_meth}{CH4 method ('leach2021' or 'thornhill2021')}

\item{ghg_meth}{GHG forcings from concentration ('myhre1998','meinshausen2020','leach2021ghg')}
}
\value{
initialised FaiR model
}
\description{
initialises a FaIRv2.1 model with configs from cbwgr::fair_configs_cmip and global scenario emissions from cbwgr::rcmip_fair
}
