% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{gen_fair_loo}
\alias{gen_fair_loo}
\title{gen_fair_loo}
\usage{
gen_fair_loo(
  rcmip_loo,
  tim_scen = "250mt-led",
  goblin_scen = "Sc3e",
  endyear = 2100L,
  ch4_meth = "thornhill2021",
  ghg_meth = "meinshausen2020"
)
}
\arguments{
\item{rcmip_loo}{global emissions dataset for global and cbwg scenarios produced by cbwg_to_loo_global_emissions}

\item{tim_scen}{a co2 ffi scenario name}

\item{goblin_scen}{afolu scenario name}

\item{endyear}{end year of run}

\item{ch4_meth}{ch4_method}

\item{ghg_meth}{ghg_method}
}
\value{
FaIR model initialised with global - ireland emissions
}
\description{
initialise a FaIR instance with global emissions minus a cbwg ireland emission scenario
}
