% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{gen_fair_loo_3}
\alias{gen_fair_loo_3}
\title{gen_fair_loo_3}
\usage{
gen_fair_loo_3(
  rcmip_loo,
  ffi_scen = "250mt-led",
  lulucf_scen = "L1",
  ag_scen = "S1",
  endyear = 2100L,
  ch4_meth = "thornhill2021",
  ghg_meth = "leach2021",
  keep_species = "All"
)
}
\arguments{
\item{rcmip_loo}{global emissions dataset for global and cbwg scenarios produced by cbwg_to_loo_global_emissions_3}

\item{ffi_scen}{a co2 ffi scenario name}

\item{lulucf_scen}{lulucf scenario name}

\item{ag_scen}{ag scenario name}

\item{endyear}{end year of run}

\item{ch4_meth}{ch4_method}

\item{ghg_meth}{ghg_method}

\item{keep_species}{names of drivers to keep, default "All"}
}
\value{
FaIR model initialised with global - ireland emissions
}
\description{
initialise a FaIR instance with global emissions minus a cbwg ireland emission scenario
}
