% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{get_conc}
\alias{get_conc}
\title{get_conc}
\usage{
get_conc(fmod, scen)
}
\arguments{
\item{fmod}{FaIR model run}

\item{scen}{a global ssp scenario}
}
\value{
a dataframe with scenario, config, specie, value and unit columns
}
\description{
get_conc
}
