% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{get_emissions}
\alias{get_emissions}
\title{get_emissions}
\usage{
get_emissions(fmod, scen)
}
\arguments{
\item{fmod}{FaIR model}

\item{scen}{global scenario}
}
\value{
a dataframe with scenario, config, specie, value and unit columns
}
\description{
get_emissions
}
