% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{get_forcing}
\alias{get_forcing}
\title{get_forcing}
\usage{
get_forcing(fmod, scen)
}
\arguments{
\item{fmod}{FaIR model run}

\item{scen}{global scenario}
}
\value{
a dataframe with scenario, config, specie, value and unit columns
}
\description{
get forcing by config and species from model run
}
