% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{get_ohc}
\alias{get_ohc}
\title{get_ohc}
\usage{
get_ohc(fmod, scen)
}
\arguments{
\item{fmod}{FaIR model}

\item{scen}{global scenario}
}
\value{
a dataframe with scenario, config, value and unit columns
}
\description{
get_ohc
}
