% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{get_warming}
\alias{get_warming}
\title{get_warming}
\usage{
get_warming(fmod, scen, use_1851_1900_baseline = TRUE)
}
\arguments{
\item{fmod}{Fair model run}

\item{scen}{global scenario}

\item{use_1851_1900_baseline}{If TRUE, use the 1851-1900 GSAT baseline}
}
\value{
a dataframe with scenario, config, value and unit columns
}
\description{
get_warming
}
