% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_scenarios.R
\docType{data}
\name{global_scenarios}
\alias{global_scenarios}
\title{Global scenarios}
\format{
A vector with 4 elements:
\describe{
  \item{1}{ssp119}
  \item{2}{ssp126}
  \item{3}{ssp434}
  \item{4}{ssp245}
}
}
\usage{
global_scenarios
}
\description{
Global scenarios
}
\examples{
global_scenarios

}
\keyword{datasets}
