% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{import_python_modules}
\alias{import_python_modules}
\title{import_python_modules}
\usage{
import_python_modules()
}
\value{
Null
}
\description{
Imports fair, Xarray, Numpy modules which need to be installed as python libraries in the python version used by
reticulate
}
