% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulucf_scenarios.R
\docType{data}
\name{lulucf_scenarios}
\alias{lulucf_scenarios}
\title{lulucf scenarios}
\format{
A vector with 4 elements:
\describe{
  \item{1}{L1}
  \item{2}{L4}
  \item{3}{L2}
  \item{4}{L3}
}
}
\usage{
lulucf_scenarios
}
\description{
lulucf scenarios considered by cbwg (ordered from weakest to strongest mitigation)
}
\examples{
lulucf_scenarios
}
\keyword{datasets}
