% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{modify_emissions}
\alias{modify_emissions}
\title{modify_emissions}
\usage{
modify_emissions(
  fmod,
  rcmip_exie,
  tim_scen = "250mt-led",
  goblin_scen = "Sc2d"
)
}
\arguments{
\item{fmod}{a FaiR model instance produced by gen_fair()}

\item{rcmip_exie}{global emissions dataset for global and cbwg scenariosn}

\item{tim_scen}{a co2 ffi scenario name}

\item{goblin_scen}{afolu scenario name}
}
\value{
FaIR model initialised with global-ireland emissions
}
\description{
initialise a FaIR instance with global emissions minus a cbwg ireland emission scenario
}
