% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tim_scenarios.R
\docType{data}
\name{tim_scenarios}
\alias{tim_scenarios}
\title{tim scenarios}
\format{
A vector with 23 elements:
\describe{
  \item{1}{450mt}
  \item{2}{450mt-wam}
  \item{3}{450mt-wem}
  \item{4}{400mt-lowbio}
  \item{5}{400mt}
  \item{6}{400mt-led}
  \item{7}{400mt-wam}
  \item{8}{400mt-wem}
  \item{9}{350mt-lowbio}
  \item{10}{350mt}
  \item{11}{350mt-led}
  \item{12}{350mt-wam}
  \item{13}{350mt-wem}
  \item{14}{300mt-wem}
  \item{15}{250mt-wem}
  \item{16}{300mt-wam}
  \item{17}{250mt-wam}
  \item{18}{300mt-lowbio}
  \item{19}{300mt-led}
  \item{20}{300mt}
  \item{21}{250mt-lowbio}
  \item{22}{250mt}
  \item{23}{250mt-led}
}
}
\usage{
tim_scenarios
}
\description{
ffi scenarios considered by cbwg
}
\examples{
tim_scenarios
}
\keyword{datasets}
