% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_model.R
\name{cbwg_to_loo_global_emissions}
\alias{cbwg_to_loo_global_emissions}
\title{cbwg_to_loo_global_emissions}
\usage{
cbwg_to_loo_global_emissions(
  global_scenarios0 = global_scenarios,
  tim_scenarios0 = tim_scenarios,
  lulucf_scenarios0 = lulucf_scenarios,
  ag_scenarios0 = ag_scenarios
)
}
\arguments{
\item{global_scenarios0}{vector of SSPs}

\item{tim_scenarios0}{vector of Energy system scenarios}

\item{lulucf_scenarios0}{vector LULUCF scenarios}

\item{ag_scenarios0}{vector AG scenarios}
}
\value{
emissions
}
\description{
Creates a global emissions dataset for specified global and Irish emissions scenarios in FaIR format. Contains 21 climate forcers.
}
